Add location mapping for external packages
When we use `@ZodImport` to reference an external package, the name
of the package is not necessarily enough to resolve it. For example,
we could be resolving it to a local location, a GitHub repo, or
other.

Now, a mapping can be configured from the package name to a resolvable
location via the Gradle plugin settings.