Unwrap value classes
Unwrap value classes. By default this is turned on, as that would
normally be the expected behavior for TypeScript schemas with value
class types. However, configuration can be used to restore the old
behavior.

A future improvement could allow specifying the unwrap option with an
annotation on a property and/or on the value class itself, if users
need to do control this on per-value-class or per-property basis.