Add location mapping for external packages (#26)
* Add location mapping for external packages

When we use `@ZodImport` to reference an external package, the name
of the package is not necessarily enough to resolve it. For example,
we could be resolving it to a local location, a GitHub repo, or
other.

Now, a mapping can be configured from the package name to a resolvable
location via the Gradle plugin settings.

* Support custom external package install commands

External package install commands are useful for overriding the
default behavior -- for example, yalc add vs npm install.

* Allow executing additional npm commands

* Install command: full command including package

Being able to customize the full command allows us to specify other things
such as the package version.

---------

Co-authored-by: Nathan Fallet <contact@nathanfallet.me>